requirements.txt
aiogram
python-dotenv
requests
googletrans